feat: Add validate basic decorator
